server-deprecated/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
aiohttp==3.9.1
prometheus-client==0.19.0
python-json-logger==2.0.7
tenacity==8.2.3

redis==5.0.1
aiohttp==3.9.1
prometheus-client==0.19.0
python-json-logger==2.0.7
tenacity==8.2.3